Add header component and npm configurations
This commit adds a new svelte component `header.svelte` for controlling the navigation. Multiple links for navigation are added in this header. It also includes a toggle button for changing the theme.

The commit introduces new npm configuration files `.npmrc` and `pnpm-lock.yaml`.
The `.npmrc` file is set to enforce engine compatibility and ensure package resolutions are of the highest version.
The introduced `pnpm-lock.yaml` provides details about the versions and integrity of all the project's dependencies.

Introduced configurations ensure consistent development environments across different devices and setups for the entire development team.