Refactored components and updated blog posts
Moved several components (Footer, Toggle, Header, and Transition) to a custom components folder. This change improves the organization of our source code, making it easier to navigate and manage these components. Updated the blog posts filenames and content, and deleted an obsolete blog post. The post filename and content are now more descriptive, which improves SEO and user experience.